drive: fix cached root_folder_id causing team_drive to be ignored.
Before this fix rclone v1.51 and 1.52 would incorrectly use the cached
root_folder_id when team_drive was set. This means that team_drive
variable was being ignored in this case.

This affects users that run rclone with different --drive-team-drive
flags.

This fix makes rclone prefer team_drive flag over root_folder_id and
issue a NOTICE message that it is ignoring the cached root_folder_id.

It also stops rclone caching the root_folder_id when using team_drive
and when team_drive is set in by the configurator it clears the
root_folder_id at the same time.

Fixes #4051